Prepare for the inevitable need for initial state
I know we'll eventually want to have urls that we can give to people
that will already have the right package loaded. I'm not sure how to
reference the current url with react, but when the time comes to figure
that out, this makes it possible to actually set that state.