PFDR-279 - Add status page
This is a rudimentary, but pretty nice status report. It shows the queue
items visible to the user in a tabbed interface with counts of failed,
pending, and completed deposits. It could be improved by including some
filtering and expand/collapse functionality for the error messages, but
this should do pretty nicely for the average case.